+++ Runtime/DB/DBManager.cs
@@ -1,9 +1,11 @@
 using System;
+using System.Linq;
 using System.Collections.Generic;
 using OcUtility;
 using Sirenix.OdinInspector;
 using UnityEditor;
 using UnityEngine;
+using Random = UnityEngine.Random;
 
 namespace OcDialogue.DB
 {
@@ -41,6 +43,21 @@ namespace OcDialogue.DB
         {
             _runtimeInitialized = false;
             Application.quitting -= Release;
+        }
+        
+        /// <summary>
+        /// (Editor Only) OcData에 사용할 id를 생성함.
+        /// </summary>
+        /// <returns></returns>
+        public static int GenerateID()
+        {
+            while (true)
+            {
+                var id = Random.Range(int.MinValue, int.MaxValue);
+                if (_instance == null) return id;
+                if (_instance.DBs.Count == 0) return id;
+                if (_instance.DBs.All(db => db.AllData.All(x => x.id != id))) return id;
+            }
         }
 #endif